requirements.txt
streamlit
ollama
fpdf2
